+++ src/main/routes/subtitle/subtitle-routes.ts
@@ -2,7 +2,7 @@ import { Request, Response, Router } from 'express'
 import { Auth } from 'googleapis'
 import env from '../../config/env'
 import { adaptRoute } from '../../adapters/express-route-adapter'
-import { makeCreateSubtitleController } from '../../factories/subtitle'
+import { makeCreateSubtitleController } from '../../factories/create-subtitle'
 import { OAuth2Client } from 'google-auth-library'
 
 const inputImage = env.inputImage
